pyproject.toml
[tool.poetry]
name = "market-access-api"
version = "0.51.00"
description = "Market Access API"
authors = [
    "Ferenc Csepregi <feri@31.digital>",
    "James Pacileo <jamespaxi@gmail.com>",
    "Ciaran Dohety <cad106uk@gmail.com>",
]

[tool.poetry.dependencies]
python = "^3.9"
boto3 = "~=1.17.86"
celery = "5.3.6"
dj-database-url = "==0.5.0"
Django = "4.2.17"
django-audit-log-middleware = "0.0.4"
django-celery-beat = "2.5.0"
django-environ = "==0.4.5"
django-extensions = "==3.0.9"
django-filter = "==2.4.0"
django-log-formatter-ecs = "==0.0.5"
django-model-utils = "==4.0.0"
django-oauth-toolkit = "==2.3.0"
django-ordered-model = "==3.4.1"
django-pglocks = "==1.0.4"
django-redis = "==4.12.1"
django-simple-history = "==3.4.0"
django-staff-sso-client = "==4.2.1"
djangorestframework = "==3.15.2"
elastic-apm = "~=6.3.0"
Faker = "~=19.3.1"
factory-boy = "~=3.0.1"
gevent = "24.2.1"
gunicorn = "==22.0.0"
hawkrest = "==1.0.1"
importlib-metadata = "~6.5"
notifications-python-client = "==6.0.2"
psycopg2-binary = "==2.9.5"
python-json-logger = "==2.0.0"
raven = "==6.9.0"
requests-toolbelt = "==0.9.1"
sentry-sdk = "2.8.0"
URLObject = "==2.4.3"
whitenoise = "==5.2.0"
django-hashid-field = "^3.1.3"
urllib3 = "~=1.26.19"
time-machine = "^2.8.2"
pytest = "^7.2.0"
certifi = "2024.7.4"
sqlparse = "0.5.0"
cryptography = "43.0.1"
tqdm = "*"
numpy = "1.24.4"
scikit-learn = "*"
scipy = "*"
nltk = "3.9.1"
sentencepiece = "*"
huggingface-hub = ">=0.4.0"
opentelemetry-distro = "0.43b0"
opentelemetry-exporter-otlp = "1.22.0"
opentelemetry-instrumentation-wsgi = "0.43b0"
opentelemetry-propagator-aws-xray = "^1.0.1"
opentelemetry-sdk-extension-aws = "^2.0.1"
django-log-formatter-asim = "^0.0.4"
dbt-copilot-python = "^0.2.0"
pandas = "2.0.0"

[tool.poetry.dev-dependencies]
black = "~=24.3.0"
boto3-stubs = "~=1.17.86"
flake8 = "~=3.8.2"
freezegun = "~=1.2.2"
ipython = "8.10.0"
mock = "~=4.0.1"
moto = "==2.2.16"
pylint-django = "~=2.3.0"
pytest-django = "~=3.10.0"
pytest-xdist = "~=2.1.0"
requests-mock = "~=1.8.0"
jsondiff = "==1.1.2"
simplejson = "==3.17.2"
ipdb = "^0.13.9"
pytest-watch = "^4.2.0"
time-machine = "^2.8.2"
greenlet = "==3.0.3"
coverage = "^7.2.7"
pytest-circleci-parallelized = "^0.1.0"
pillow = "10.3.0"
# Below are the packages required for the ML models, we keep them in dev-dependencies so they are not put into the
# requirements.txt file and installed in production - trying to do so causes ALL sorts of issues.
# instead we install them as part of the .profile script that is ran on the server during deployment.
transformers = "4.38.0"
sentence-transformers = "3.0.1"
torch = "2.2.0"
torchvision = "0.17.0"


[build-system]
requires = ["poetry~=1.6.1"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
testpaths = "tests"
python_files = ["tests.py", "test_*.py", "*_tests.py"]

addopts = """
    --reuse-db
    --ds=config.settings.test
"""

[tool.coverage.run]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*__init__*",
    "*test_*",
]
include = [
    "api/*",
]
branch = true

[tool.isort]
profile = "black"
